Got the pool memory manager working. Not nearly as significant of a performance boost as it was when the Efficient C++ book was written.
This further confirms my beliefs that compiler technology has
come a long way since the book was written.

Sadly it is starting to seem to me that there isn't nearly as
much of an importance on human optimization anymore...
While this is good in general, I miss trying to squeeze out
as much performance as I can...

That all being said the pool memory allocator definitely does
a better job than global allocator. Just nowhere near an
order of magnitude better

On the alloc_test code.
global: ~1s
pool: ~0.8s